Support oneway GPIO pins that are auto exported
This is needed to support GPIOs added by the gpio_helper
driver on the Beaglebone. These GPIOs come from a device
tree description so Linux automatically knows to export
them. The direction is specified in the dtb file, so it
can't be changed, and in fact, the direction file doesn't
even exist.